Fix crashes in eval_pair()
size was initialized at 0 so when it was doubled to grow the new list it
tried to reallocate the memory to 0 bytes. The realloc was also trying
to change the passed the argument not the new list. Additionally cap and
size were reversed adding addition confusing. Who wrote this shit, oh
wait.