Compartmentalize function scopes with linked hashs
Fixes various issues with macros and functions using old overlapping
variables from different scopes. Each time a function or macro is run a
new environment hash is created to hold that procedure's arguments,
which then links to the old hash. The last hash in this list is the
global namespace where define creates new variables. Variable look up
goes through all hashes start with the most recently created one. At the
end of a function this new hash is simply removed from the front of the
list.