Add line reading syntax to replace s-expressions
Each new line is read as a list, with implied parenthesis around it.
If a line only has one element then return just that value, not a list.
A new line that is indented more than the previous line is a sub
expression of the previous line's list. It does not matter how many
tabs or spaces are in each level of indent as long as it is the same
number as other lines on the same level. For example:

```
def fact[n]
  if (= n 1)
    1
    * n fact[(- n 1)]
```

Is equivalent to:

```
(def fact[n]
  (if (= n 1)
    1
    (* n fact[(- n 1)])))
```